backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.1

# Google Gemini AI
google-generativeai==0.3.1

# Image Generation
diffusers==0.25.0
torch==2.2.0
transformers==4.37.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
httpx==0.26.0

# Optional - for development
black==24.1.1
isort==5.13.2
flake8==7.0.0
